# [PERF] tokenizer: avoid regexp match